requirements.txt
pandas
scikit-learn
matplotlib
typer
seaborn
pyyaml